Split entity browser components into multiple files
